Emulator code updates
* Make major changes to emulator and fix several bugs from its first
draft.
* Change cyclesRemaining and elapsedTicks from unsigned integer to
signed. This addresses a loop check error where elapsedTicks is expected
to be less than zero.
* Remove macros for getting address values and memory locations. Using
macros made it too hard to debug well.
* Add functions for getting address values and memory locations.
Functions will be inlined by a good compiler and now debugging is a lot
easier. Changes also include several incorrect retrieval methods from
the first draft (primarily accessing indirect memory incorrectly.
* Add GetImmediate8 and SetImmediate8, which first check if the memory
location being accessed is a register. If it is, the call is redirected
to a register-specific function.
* Fix SetCarry, ClearCarry, and CheckCarry mathods to call the CMask
value instead of the bit index.
* Refactor CreateReadInstructions, CreateWriteInstructions, and
CreateRefInstructions to call the new memory access functions.
* Remove M2X and X2M arrays, which were a byproduct of the x86 assembly
code and not needed for C.
* Add OpSize array, which provides the number of bytes for each
instruction, minus one.
* Fix PushWord function to push a word onto the stack in the right
order.
* Fix Pull function to retrieve a value from the stack at the right
index.
* Fix PullWord function to pull a word from the stack in the right
order.
* Add LDN function, which loads a value into a given register.
* Refactor LD macro to call LDN function.
* Update INC, DEC, LSR, ASL, ROL, and ROR functions to call
GetImmediate8 and SetImmediate8 for non-registers.
* Fix Compare method to write properly to the Z flag.
* Have Branch function properly go to the next instruction if a branch
doesn't happen.
* Create direct JMP functions, since their memory calls are different
from what the CreateReadInstructions macro would build for them,
* Update JSR function to push the right PC value.
* Add direct JSR function since there is only one, and it is different
from what the CreateReadInstructions macro would create for it.
* Add Transfer functions, for the TAX, TAY, TXA, etc. methods.
* Add GetWord functions, which gets the 16-bit value at a specified
address.
* Update PC update from vector interrupts to get correct 16-bit address.
* Remove usage of AltFlags variable.
* Remove undefining of ragister macros. They exist only in the scope of
the M6502.c file so undefining them is unneccesary.
* Refactor IndirectIndex to IndirectY and IndexedIndirect to IndirectX,
to make more clear which is which.
* Switch opcodes calling IndirectY and IndirectX, all of which were
calling the wrong one before.