Add missing includes to several header files
This is exactly why local includes should come before standard library.
It really helps highlight what is contained in a every compilation unit
and header. And you know what I see? The includes and data types here
are a damn mess. So many of these structures should not be based on
Windows integral typedefs. That's such a proprietary header to include.
I'll need to make stdint save the day soon.