Remove outdated _WIN32_IE define
This define was an old Win32 guard to ensure that the user has an
updated comctl32.dll on their system (which is needed here for several
controls and buttons that we use). Seeing as Internet Explorer is
long-since retired, and any comctrl32.dll in the last 15 years would be
of sufficient requirement, I would say that it is safe to remove this
old guard.

If I REALLY wanted to check for some minimum required version, I could
use NTDDI_VERSION (see sdkdkver.h).